requirements.txt
wheel>=0.22
pytest
biopython
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
pysam